+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [4.5.1](https://github.com/LabShare/services/compare/v4.5.0...v4.5.1) (2020-04-09)
+
+
+### Bug Fixes
+
+* path parameters get lower-cased ([71ed923](https://github.com/LabShare/services/commit/71ed92319ea93b4fe61efe0fa5208caf40ba84f4))
+
+# [4.5.0](https://github.com/LabShare/services/compare/v4.4.0...v4.5.0) (2020-04-08)
+
+
+### Features
+
+* added options to disable authentication and set user info ([6a017b5](https://github.com/LabShare/services/commit/6a017b5c6952cb7e1c0d0cfc117be585e345f9c7))
+
 # [4.4.0](https://github.com/LabShare/services/compare/v4.3.0...v4.4.0) (2020-04-01)